// models/survey-selectee.ts
export class SurveySelectee {
  SurveySelecteeId: number;
  SurveyId: number;
  EmployeeId: number;
  FirstName: string;
  LastName: string;
  Email: string;
  Role: string;
  Department: string;
  CompanyId: string;
  Miscellaneous: string;
  IsDeleted: boolean;
}

export class SurveySelecteeIds {
  SurveySelecteeId: number[];
}